requirements.txt
asgiref==3.5.0
Django==4.0.3
sqlparse==0.4.2
django-ckeditor==6.7.0
djangorestframework==3.14.0
drf-spectacular==0.26.5